chore(scripts): Move scripts direcotry one level above
