+++ scripts/local-network.js
@@ -1,6 +1,6 @@
 const TruffleContract = require('truffle-contract');
 const Web3 = require('web3');
-const config = require('./config');
+const config = require('../src/config');
 const WTIndexContract = require('@windingtree/wt-contracts/build/contracts/WTIndex');
 
 const provider = new Web3.providers.HttpProvider(config.get('web3Provider'));
