+++ src/index.js
@@ -1,6 +1,6 @@
 const { app } = require('./app');
 const config = require('./config');
-const { deployIndex } = require('./scripts/local-network');
+const { deployIndex } = require('../scripts/local-network');
 
 const server = app.listen(config.get('port'), () => {
   if (process.env.ETH_NETWORK === 'local') {
